feat: Add function to change base image in IapBuilder
This commit adds a new function `change_base_image` to the `IapBuilder` class. This function allows users to change the base image used in the IAP (Identity Access Proxy) configuration. It makes a POST request to the IAP API with the provided image URL and updates the base image for the specified company.

The `change_base_image` function is called in the code to update the base image after creating an edge app using the `IapBuilder` class.

This change improves the flexibility and customization options for IAP configurations.